+++ src/interfaces/TypeResolver.ts
@@ -1,10 +1,8 @@
-import { GraphQLResolveInfo } from "graphql";
+import { GraphQLTypeResolver } from "graphql";
 
 import { ClassType } from "./ClassType";
 import { MaybePromise, Maybe } from "./Maybe";
 
 export type TypeResolver<TSource, TContext> = (
-  value: TSource,
-  context: TContext,
-  info: GraphQLResolveInfo,
+  ...args: Parameters<GraphQLTypeResolver<TSource, TContext>>
 ) => MaybePromise<Maybe<string | ClassType>>;